Remove unnecessary C libraries for db-mysql.
In particular, this fixes db-mysql for Windows, for which the dlls
don't work. (32 bit doesn't work, and 64 bit doesn't exist. The
Makefiles are literally decades old)

The C library isn't really required. The history of the file goes back
to the initial commit, so it's hard to figure out the exact context in
which it was required. It seems to have been used for two things:

1. Returning 64 bit values: Perhaps at the time some Lisp
implementation didn't support 64 bit in the FFI. I'm pretty confident
all the CLs have 64-bit support in FFIs these days, so we don't need
the C library for this purpose, and just use the direct mysql
functions while returning :unsigned-long-long type.

2. Accesssing structure of MYSQL_FIELD. The comments say this is so
that we don't have to depend on mysql version (the C libary, if it's
built against the right mysql.h, should automatically get this
struct.) This makes sense, but this struct hasn't changed since Mysql
4.1, so we can assume that's it's more or less stabilized.

Finally, there were a bunch of unused functions that could easily be
removed.

With this change, I'm finally able to use CLSQL+MySQL under Windows!
So it fixes, for instance https://github.com/UnwashedMeme/clsql/issues/9

I ran tests on Lispworks 7.1.2 and SBCL running under Debian 10
64-bit. I didn't run tests under Windows, but I'm using it in my app
and it's working find on a Windows laptop. (PS. CLSQL needs some other
changes to work under LW7, I'll send a separate PR for that.)